change in message class

+++ constantes/constantes.go
@@ -3,7 +3,7 @@ package constantes
 const MulticastAddr = "224.0.0.1:6666"
 const ListeningPort = ":6667"
 
-const AttenteK uint8 = 3
+const AttenteK int = 3
 
 const SYNC uint8 = 0
 const FOLLOW_UP uint8 = 1
